+++ examples/resources/iosxr_interface_ethernet/resource.tf
@@ -1,7 +1,6 @@
 resource "iosxr_interface_ethernet" "example" {
   type                         = "GigabitEthernet"
   name                         = "0/0/0/1"
-  l2transport                  = false
   point_to_point               = false
   multipoint                   = false
   dampening                    = true