+++ internal/provider/data_source_iosxr_interface_bundle_ether_subinterface_test.go
@@ -33,6 +33,7 @@ import (
 
 func TestAccDataSourceIosxrInterfaceBundleEtherSubinterface(t *testing.T) {
 	var checks []resource.TestCheckFunc
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "l2transport", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "point_to_point", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "multipoint", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "dampening", "true"))
@@ -197,6 +198,14 @@ func TestAccDataSourceIosxrInterfaceBundleEtherSubinterface(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "lldp", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "lldp_transmit_disable", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "lldp_receive_disable", "true"))
+	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "monitor_sessions.0.session_name", "SESSION-1"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "monitor_sessions.0.ethernet", "true"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "monitor_sessions.0.direction_rx_only", "true"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "monitor_sessions.0.acl", "true"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "monitor_sessions.0.acl_ipv4_name", "ACL1"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "monitor_sessions.0.acl_ipv6_name", "ACL2"))
+	}
 	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
 		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether_subinterface.test", "ptp", "true"))
 	}
@@ -476,6 +485,14 @@ resource "iosxr_gnmi" "PreReq4" {
 	]
 }
 
+resource "iosxr_gnmi" "PreReq5" {
+	path = "Cisco-IOS-XR-um-monitor-session-cfg:/monitor-sessions/monitor-session[session-name=SESSION-1]"
+	attributes = {
+		"session-name" = "SESSION-1"
+		"traffic-type" = "ethernet"
+	}
+}
+
 `
 
 // End of section. //template:end testPrerequisites
@@ -486,6 +503,7 @@ func testAccDataSourceIosxrInterfaceBundleEtherSubinterfaceConfig() string {
 	config := `resource "iosxr_interface_bundle_ether_subinterface" "test" {` + "\n"
 	config += `	delete_mode = "attributes"` + "\n"
 	config += `	name = "100.100"` + "\n"
+	config += `	l2transport = false` + "\n"
 	config += `	point_to_point = false` + "\n"
 	config += `	multipoint = false` + "\n"
 	config += `	dampening = true` + "\n"
@@ -672,6 +690,16 @@ func testAccDataSourceIosxrInterfaceBundleEtherSubinterfaceConfig() string {
 	config += `	lldp = true` + "\n"
 	config += `	lldp_transmit_disable = true` + "\n"
 	config += `	lldp_receive_disable = true` + "\n"
+	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
+		config += `	monitor_sessions = [{` + "\n"
+		config += `		session_name = "SESSION-1"` + "\n"
+		config += `		ethernet = true` + "\n"
+		config += `		direction_rx_only = true` + "\n"
+		config += `		acl = true` + "\n"
+		config += `		acl_ipv4_name = "ACL1"` + "\n"
+		config += `		acl_ipv6_name = "ACL2"` + "\n"
+		config += `	}]` + "\n"
+	}
 	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
 		config += `	ptp = true` + "\n"
 	}
@@ -861,7 +889,7 @@ func testAccDataSourceIosxrInterfaceBundleEtherSubinterfaceConfig() string {
 		config += `		clock_class_to_map_to = 6` + "\n"
 		config += `	}]` + "\n"
 	}
-	config += `	depends_on = [iosxr_gnmi.PreReq0, iosxr_gnmi.PreReq1, iosxr_gnmi.PreReq2, iosxr_gnmi.PreReq3, iosxr_gnmi.PreReq4, ]` + "\n"
+	config += `	depends_on = [iosxr_gnmi.PreReq0, iosxr_gnmi.PreReq1, iosxr_gnmi.PreReq2, iosxr_gnmi.PreReq3, iosxr_gnmi.PreReq4, iosxr_gnmi.PreReq5, ]` + "\n"
 	config += `}` + "\n"
 
 	config += `