+++ internal/provider/data_source_iosxr_interface_bundle_ether_test.go
@@ -230,6 +230,15 @@ func TestAccDataSourceIosxrInterfaceBundleEther(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "lldp_transmit_disable", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "lldp_receive_disable", "true"))
 	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "monitor_sessions.0.session_name", "SESSION-1"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "monitor_sessions.0.ethernet", "true"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "monitor_sessions.0.direction_rx_only", "true"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "monitor_sessions.0.port_level", "true"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "monitor_sessions.0.acl", "true"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "monitor_sessions.0.acl_ipv4_name", "ACL1"))
+		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "monitor_sessions.0.acl_ipv6_name", "ACL2"))
+	}
+	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
 		checks = append(checks, resource.TestCheckResourceAttr("data.iosxr_interface_bundle_ether.test", "ptp", "true"))
 	}
 	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
@@ -500,6 +509,14 @@ resource "iosxr_gnmi" "PreReq3" {
 	]
 }
 
+resource "iosxr_gnmi" "PreReq4" {
+	path = "Cisco-IOS-XR-um-monitor-session-cfg:/monitor-sessions/monitor-session[session-name=SESSION-1]"
+	attributes = {
+		"session-name" = "SESSION-1"
+		"traffic-type" = "ethernet"
+	}
+}
+
 `
 
 // End of section. //template:end testPrerequisites
@@ -729,6 +746,17 @@ func testAccDataSourceIosxrInterfaceBundleEtherConfig() string {
 	config += `	lldp_transmit_disable = true` + "\n"
 	config += `	lldp_receive_disable = true` + "\n"
 	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
+		config += `	monitor_sessions = [{` + "\n"
+		config += `		session_name = "SESSION-1"` + "\n"
+		config += `		ethernet = true` + "\n"
+		config += `		direction_rx_only = true` + "\n"
+		config += `		port_level = true` + "\n"
+		config += `		acl = true` + "\n"
+		config += `		acl_ipv4_name = "ACL1"` + "\n"
+		config += `		acl_ipv6_name = "ACL2"` + "\n"
+		config += `	}]` + "\n"
+	}
+	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
 		config += `	ptp = true` + "\n"
 	}
 	if os.Getenv("NCS") != "" || os.Getenv("C8000") != "" {
@@ -917,7 +945,7 @@ func testAccDataSourceIosxrInterfaceBundleEtherConfig() string {
 		config += `		clock_class_to_map_to = 6` + "\n"
 		config += `	}]` + "\n"
 	}
-	config += `	depends_on = [iosxr_gnmi.PreReq0, iosxr_gnmi.PreReq1, iosxr_gnmi.PreReq2, iosxr_gnmi.PreReq3, ]` + "\n"
+	config += `	depends_on = [iosxr_gnmi.PreReq0, iosxr_gnmi.PreReq1, iosxr_gnmi.PreReq2, iosxr_gnmi.PreReq3, iosxr_gnmi.PreReq4, ]` + "\n"
 	config += `}` + "\n"
 
 	config += `