+++ internal/provider/resource_iosxr_interface_ethernet_test.go
@@ -37,7 +37,6 @@ func TestAccIosxrInterfaceEthernet(t *testing.T) {
 	var checks []resource.TestCheckFunc
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_interface_ethernet.test", "type", "GigabitEthernet"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_interface_ethernet.test", "name", "0/0/0/1"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxr_interface_ethernet.test", "l2transport", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_interface_ethernet.test", "point_to_point", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_interface_ethernet.test", "multipoint", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_interface_ethernet.test", "dampening", "true"))
@@ -567,7 +566,6 @@ func testAccIosxrInterfaceEthernetConfig_all() string {
 	config := `resource "iosxr_interface_ethernet" "test" {` + "\n"
 	config += `	type = "GigabitEthernet"` + "\n"
 	config += `	name = "0/0/0/1"` + "\n"
-	config += `	l2transport = false` + "\n"
 	config += `	point_to_point = false` + "\n"
 	config += `	multipoint = false` + "\n"
 	config += `	dampening = true` + "\n"